node-main/deps/v8/test/inspector/regress/regress-crbug-1081162-expected.txt

Regression test for crbug/1081162

Running test: testBreakpointResolved
After Debugger.setBreakpointByUrl
async function foo(o) {
  await o.#bar();
}

After Debugger.breakpointResolved
async function foo(o) {
  await o.#bar();
}
